Added auto-rotation

+++ src/hrb/BlobMeta.hh
@@ -37,14 +37,14 @@ public:
 	rapidjson::Document serialize() const;
 
 	const std::string& mime() const {return m_mime;}
-	int orientation() const {return m_orientation;}
+	long orientation() const {return m_orientation;}
 	const std::string& filename() const {return m_filename;}
 
 	void filename(std::string_view fn) {m_filename = fn;}
 
 private:
 	std::string m_mime{"application/octet-stream"};
-	int m_orientation{1};
+	long m_orientation{1};
 	std::string m_filename;
 };
